+++ tests/template.spec.ts
@@ -1,6 +1,6 @@
 import { assert } from "chai";
 
-import { CocoaBuilder, ephemeral } from "../src/template";
+import { CocoaBuilder, Ephemeral } from "../src/template";
 
 describe("Template Module", () => {
     it("Cocoa Builder", () => {
@@ -9,32 +9,36 @@ describe("Template Module", () => {
         assert.equal(b.name, "test");
     });
 
-    it("ephemeral", () => {
-        class Ephemeral {
+    it("Ephemeral", () => {
+        class DummyEphemeral {
             constructor(
                 public name?: string,
                 public description?: string,
                 public required?: boolean
             ) {}
             setName(name: string) {
-                return new Ephemeral(name, this.description, this.required);
+                return new DummyEphemeral(
+                    name,
+                    this.description,
+                    this.required
+                );
             }
             setDescription(desc: string) {
-                return new Ephemeral(this.name, desc, this.required);
+                return new DummyEphemeral(this.name, desc, this.required);
             }
             setRequired(req: boolean) {
-                return new Ephemeral(this.name, this.description, req);
+                return new DummyEphemeral(this.name, this.description, req);
             }
         }
 
-        let ep = new Ephemeral();
+        let ep = new DummyEphemeral();
 
-        ep = ephemeral()(ep);
+        ep = Ephemeral()(ep);
 
         assert.equal(ep.name, "ephemeral");
         assert.equal(ep.description, "Make your request ephemeral");
 
-        ep = ephemeral("bruh")(ep);
+        ep = Ephemeral("bruh")(ep);
 
         assert.equal(ep.description, "bruh");
         assert.strictEqual(ep.required, false);