+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notables change for Cocoa Discord Utils will be noted here
 
 **Note**: Changelog before 1.0.0-pre is not available here
 
+## [1.0.0-rc.4] - 2022-02-??
+
+- **BREAKING**: `ephemeral` is now `Ephemeral`, and some few interfaces have
+their name changed
+
+- Added `getEphemeral(ctx)` to shortly retrieve ephemeral option
+
 ## [1.0.0-rc.3] - 2022-02-09
 
 - Updated Error Handling